+++ yarn-project/pxe/src/contract_function_simulator/note_validation_request.ts
